AudioService: Show safe volume panel for non-interactive volume adjustments
If safe volume is active, apps calling mp.setVolume() to adjust the headset to a value higher than the threshold will have the operation fail silently because no volume dialog is active to pop up the "do you wish to override" question. This is particularly bad for peripherals that use the audio jack (Square readers, Jawbone bands), since they usually require levels close to maximum for proper communication and the app isn't prepared to deal with a silent rejection of the requested volume.